init server and structure pages
