--- modules/DashboardPages/HomePage.tsx
@@ -1,27 +1,0 @@
-import React from "react";
-import Image from "next/image";
-import { Project } from "../../ui";
-
-const HomePage = () => {
-  return (
-    <div className="ml-60">
-      <div className="dashboard-header">
-        <div>Home</div>
-      </div>
-      <div className="grid grid-cols-4 gap-8 overflow-auto px-7 pt-24">
-        <Project
-          title="title"
-          description="description"
-          imgSrc="https://cdn.dribbble.com/users/2624832/screenshots/15457526/media/e155407d8d3f0b64c0be314aa7ac8cc6.png"
-        />{" "}
-        <Project
-          title="title"
-          description="description"
-          imgSrc="https://cdn.dribbble.com/users/6237882/screenshots/15270492/media/11863e6e28299b753fcff57b37793e98.png"
-        />
-      </div>
-    </div>
-  );
-};
-
-export default HomePage;
