--- modules/DashboardPages/SettingsPage.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-
-const SettingsPage = () => {
-  return (
-    <div className="ml-60">
-      <div className="dashboard-header">Settings</div>
-      <div className="">SettingContent</div>
-    </div>
-  );
-};
-
-export default SettingsPage;
