+++ pages/dashboard/settings.tsx
@@ -1,3 +1,12 @@
-import SettingsPage from "../../modules/DashboardPages/SettingsPage";
+import React from "react";
+
+const SettingsPage = () => {
+  return (
+    <div className="ml-60">
+      <div className="dashboard-header">Settings</div>
+      <div className="">SettingContent</div>
+    </div>
+  );
+};
 
 export default SettingsPage;
